refactor(stream): split index.ts into per-class modules
The 1676-line packages/node/stream/src/index.ts is split into per-class
modules (Workstream E):

- stream-base.ts        Stream_ (root EventEmitter + pipe entry point)
- readable.ts           Readable_ (+ protected _autoClose hook)
- writable.ts           Writable_ (+ FIFO drain queue)
- duplex.ts             Duplex_   (Readable_ + writable-half re-impl)
- transform.ts          Transform_
- passthrough.ts        PassThrough_
- utils/pipe.ts         Stream.pipe() helper
- utils/pipeline.ts     pipeline()
- utils/finished.ts     finished(), addAbortSignal(), is{Readable,...}()
- internal/state.ts     module-singleton defaults + validateHighWaterMark
- internal/types.ts     shared interfaces

src/index.ts is now a thin re-export barrel that preserves the historical
default-export shape (Stream augmented with {Readable, Writable, Duplex,
Transform, PassThrough, pipeline, finished, addAbortSignal, isReadable,
isWritable, isDestroyed, isDisturbed, isErrored, getDefaultHighWaterMark,
setDefaultHighWaterMark}) so cjs-compat.cjs's `mod.default || mod` and
`util.inherits(Sub, require('stream'))` keep working unchanged.

Stream_.prototype.pipe is wired via a late-binding _setPipeImpl hook
rather than a direct top-level import. The natural dependency graph
(stream-base → pipe → readable → stream-base) trips GJS's eager ESM
evaluation with "class heritage Stream_ is not an object or null" — when
pipe.ts statically imports Readable_ and stream-base.ts statically
imports pipe.ts, Readable_ resolves to undefined while Stream_ is being
declared. The hook pattern lets stream-base.ts stay dependency-free of
its own subclasses; utils/pipe.ts wires itself in via a side-effect call
at module-load time, and src/index.ts adds a `import './utils/pipe.js'`
side-effect import to ensure it runs before any pipe() call.

`as any` in the source files is reduced from 35 occurrences to 0 —
replaced by typed `unknown` casts, concrete stream subtypes, or narrow
`as unknown as` bridges where Node's option-type `this: Readable`
doesn't unify with our internal `Readable_`. Two private chunkLen
helpers (writable.ts, duplex.ts) replace the inline `(chunk?.length ?? 1)`
pattern at four call sites each.

Test results unchanged from baseline:
  - @gjsify/stream: 509 GJS / 507 Node (all green)
  - @gjsify/integration-streamx: 156 GJS / 155 Node (all green)
  - @gjsify/integration-socket.io: 107/112 GJS, 112/112 Node (5 pre-
    existing GJS failures: 4× volatile-event timing + 1× disconnect-all-
    namespaces timeout — confirmed identical on baseline)
  - @gjsify/integration-webtorrent: pre-existing setup error
    `parseTorrent.toMagnetURI is not a function` from a parse-torrent
    npm-package shape change, unrelated to this refactor (confirmed
    identical on baseline)
