+++ STATUS.md
@@ -68,7 +68,7 @@ The project comprises **43 Node.js packages** (+1 meta), **19 Web API packages**
 | **process** | GLib, GjsifyTerminal | 143 (2 specs) | EventEmitter-based, env (CRUD, enumerate, coerce), cwd/chdir, platform, arch, pid/ppid, version/versions, argv, hrtime/hrtime.bigint (**monotonicity, diff**), memoryUsage (**field validation**), nextTick (**FIFO ordering, args**), exit/kill, config, execArgv, cpuUsage (**delta**), **signal handler registration**, **stdout/stderr write methods**, **emitWarning**; stdin/stdout/stderr as ProcessReadStream/ProcessWriteStream with `isTTY`, `setRawMode`, `columns`/`rows` via `@gjsify/terminal-native` (native) or env/GLib fallback; SIGWINCH→stdout/stderr 'resize' |
 | **querystring** | — | 471 | parse/stringify with full encoding |
 | **readline** | — | 145 (2 specs) | Interface (lifecycle, line events, mixed line endings, Unicode, chunked input, long lines, history), question (sequential, output), prompt, pause/resume, async iterator, clearLine/clearScreenDown/cursorTo/moveCursor, **readline/promises** (createInterface, question→Promise) |
-| **stream** | — | 509 (7 specs) | Readable, Writable, Duplex, Transform (**_flush** edge cases, constructor options, objectMode, split HWM, destroy, final/flush ordering, ERR_MULTIPLE_CALLBACK), PassThrough, objectMode, backpressure (**drain events**, **HWM=0**), **pipe** (event, cleanup, error handling, multiple dest, unpipe, same dest twice, needDrain, objectMode→non-objectMode), **inheritance** (instanceof hierarchy, util.inherits single/multi-level, stream subclassing), destroy, **pipeline** (error propagation, multi-stream), **finished** (premature close, cleanup), **addAbortSignal**, **Readable.from** (array/generator/async generator/string/Buffer), consumers (text/json/buffer/blob/arrayBuffer), promises (pipeline/finished), **async iteration**, **_readableState/_writableState** (highWaterMark, objectMode, pipes), **Symbol.hasInstance** (Duplex/Transform/PassThrough instanceof Writable) |
+| **stream** | — | 509 (7 specs) | Readable, Writable, Duplex, Transform (**_flush** edge cases, constructor options, objectMode, split HWM, destroy, final/flush ordering, ERR_MULTIPLE_CALLBACK), PassThrough, objectMode, backpressure (**drain events**, **HWM=0**), **pipe** (event, cleanup, error handling, multiple dest, unpipe, same dest twice, needDrain, objectMode→non-objectMode), **inheritance** (instanceof hierarchy, util.inherits single/multi-level, stream subclassing), destroy, **pipeline** (error propagation, multi-stream), **finished** (premature close, cleanup), **addAbortSignal**, **Readable.from** (array/generator/async generator/string/Buffer), consumers (text/json/buffer/blob/arrayBuffer), promises (pipeline/finished), **async iteration**, **_readableState/_writableState** (highWaterMark, objectMode, pipes), **Symbol.hasInstance** (Duplex/Transform/PassThrough instanceof Writable). **Source split (Workstream E):** the 1676-line `src/index.ts` is split into per-class modules — `stream-base.ts` (Stream_), `readable.ts`, `writable.ts`, `duplex.ts`, `transform.ts`, `passthrough.ts`, `utils/{pipe,pipeline,finished}.ts`, `internal/{state,types}.ts`. `src/index.ts` is now a re-export barrel preserving the `cjs-compat.cjs` default-export shape (`Stream` + classes/helpers attached). `Stream_.prototype.pipe` is wired via a late-binding hook (`_setPipeImpl`) to break the stream-base→pipe→readable→stream-base import cycle that GJS's eager ESM evaluation otherwise hits with "class heritage … is not an object or null". `as any` reduced from 35 occurrences to 0 — replaced by typed `unknown` casts and concrete stream subtypes throughout. |
 | **string_decoder** | — | 103 | UTF-8, Base64, hex, streaming |
 | **sys** | — | 7 | Deprecated Node alias — re-exports `@gjsify/util` |
 | **timers** | — | 88 (3 specs) | setTimeout/setInterval/setImmediate (**delay verification, args, clear, ordering**) + timers/promises |
